:construction: Add pager current item rx extensions
